new custom font renderer

+++ Firework-Client/src/main/java/com/firework/client/Implementations/Managers/Text/TextManager.java
@@ -26,7 +26,7 @@ public class TextManager {
 
     public int getStringWidth(String text) {
         if(customFont){
-            return round(Firework.customFontManager.getWidth(text));
+            return (int) round(Firework.customFontManager.getWidth(text));
         }else {
             return mc.fontRenderer.getStringWidth(text);
         }
